Added parent guide plus cleanup
Added a new "Current students" section and moved the volunteer/board
position info under that. Created a new documents page to hold the
parent handbook(s). Added the absentee line info to the "Current
students" page. Preliminary French translation with assistance from
Google for the new content. Removed references to the 2014 open house
and 2014 registration in general. Updated the node modules to their
latest versions, got rid of modules we're not currently using.